Add progress reporting to reindexing

+++ libraries/blockchain/include/bts/blockchain/chain_database.hpp
@@ -96,8 +96,8 @@ namespace bts { namespace blockchain {
           * @brief open Open the databases, reindexing as necessary
           * @param reindex_status_callback Called for each reindexed block, with the count of blocks reindexed so far
           */
-         void open(const fc::path& data_dir, fc::optional<fc::path> genesis_file ,
-                   std::function<void(uint32_t)> reindex_status_callback = std::function<void(uint32_t)>());
+         void open(const fc::path& data_dir, fc::optional<fc::path> genesis_file,
+                   std::function<void(float)> reindex_status_callback = std::function<void(float)>());
          void close();
 
          void add_observer( chain_observer* observer );
